Back-End/pyproject.toml
[project]
name = "Proyecto BD2025"
version = "0.1.0"
description = "Proyecto para gestionar Marloy Coffee"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.1",
    "fastapi>=0.115.8",
    "pydantic>=2.10.6",
    "requests>=2.32.3",
    "sqlmodel>=0.0.22",
    "uvicorn>=0.34.0",
    "mysql-connector-python>=8.3.0",
    "cryptography>=42.0.0"
]
